pyproject.toml
[tool.poetry]
name = 'feedo'
version = '0.1.0'
description = 'Simple RSS Feed backend'

authors = ['Andrey Gordeev <gordeev.and.and@gmail.com>']

readme = 'README.md'
include = ['CHANGELOG']

[tool.poetry.dependencies]
python = '^3.6'
django = '^2'
celery = '^4.4.0'
django_celery_beat = '^1.5.0'
jinja2 = '^2.10.3'
feedparser = '^5.2.1'
psycopg2 = "==2.8.4"

[tool.poetry.dev-dependencies]
django-configurations = '^2.1'
pytest-django = '^3.4'
pytest-pythonpath = '^0.7.3'
mock = '^1.0.1'
freezegun = '^0.3.7'

[build-system]
requires = ['poetry>=0.12']
build-backend = 'poetry.masonry.api'
